requirements.txt
asgiref==3.5.2
Django==4.0.8
psycopg2-binary==2.9.3
djangorestframework
requests
django-phonenumber-field[phonenumberslite]
django-htmx==1.6.0